Divided Main manager class into variety of focused services (#223)
