+++ ui/src/features/Map/index.tsx
@@ -16,8 +16,15 @@ import { sourceConfigs, SourceId } from '@/features/Map/sources';
 import { drawnFeatureContainsExtent, getSelectedColor, getSortKey } from '@/features/Map/utils';
 import { showGraphPopup } from '@/features/Popup/utils';
 import { useSpatialSelection } from '@/hooks/useSpatialSelection';
-import mainManager from '@/managers/Main.init';
 import notificationManager from '@/managers/Notification.init';
+import {
+  configService,
+  factoryService,
+  locationService,
+  mainManager,
+  mapService,
+  popupService,
+} from '@/services/init';
 import useMainStore from '@/stores/main';
 import { Location } from '@/stores/main/types';
 import useSessionStore from '@/stores/session';
@@ -91,6 +98,8 @@ const MainMap: React.FC<Props> = (props) => {
     if (initialMapLoad.current) {
       initialMapLoad.current = false;
       mainManager.setMap(map);
+      mapService.setMap(map);
+      configService.setMap(map);
       map.resize();
       map.fitBounds(
         [
@@ -114,7 +123,7 @@ const MainMap: React.FC<Props> = (props) => {
       return;
     }
 
-    mainManager.setHoverPopup(hoverPopup);
+    popupService.setHoverPopup(hoverPopup);
   }, [hoverPopup]);
 
   useEffect(() => {
@@ -122,7 +131,7 @@ const MainMap: React.FC<Props> = (props) => {
       return;
     }
 
-    mainManager.setPersistentPopup(persistentPopup);
+    popupService.setPersistentPopup(persistentPopup);
   }, [persistentPopup]);
 
   useEffect(() => {
@@ -130,7 +139,7 @@ const MainMap: React.FC<Props> = (props) => {
       return;
     }
 
-    mainManager.setContainer(container);
+    popupService.setContainer(container);
   }, [container]);
 
   useEffect(() => {
@@ -138,7 +147,8 @@ const MainMap: React.FC<Props> = (props) => {
       return;
     }
 
-    mainManager.setDraw(draw);
+    mapService.setDraw(draw);
+    configService.setDraw(draw);
   }, [draw]);
 
   useEffect(() => {
@@ -160,7 +170,7 @@ const MainMap: React.FC<Props> = (props) => {
 
     const allIds: string[] = [];
     layers.forEach((layer) => {
-      const { pointLayerId, lineLayerId, fillLayerId } = mainManager.getLocationsLayerIds(
+      const { pointLayerId, lineLayerId, fillLayerId } = factoryService.getLocationsLayerIds(
         layer.datasourceId,
         layer.id
       );
@@ -186,7 +196,7 @@ const MainMap: React.FC<Props> = (props) => {
           if (features && features.length > 0) {
             hoverPopup.remove();
 
-            const uniqueFeatures = mainManager.getUniqueIds(features, layer.id);
+            const uniqueFeatures = locationService.getUniqueIds(features, layer.id);
             const locations: Location[] = uniqueFeatures.map(({ id }) => ({
               id,
               layerId: layer.id,
@@ -212,7 +222,7 @@ const MainMap: React.FC<Props> = (props) => {
     }
 
     layers.forEach((layer) => {
-      const { pointLayerId, lineLayerId, fillLayerId } = mainManager.getLocationsLayerIds(
+      const { pointLayerId, lineLayerId, fillLayerId } = factoryService.getLocationsLayerIds(
         layer.datasourceId,
         layer.id
       );
@@ -245,7 +255,7 @@ const MainMap: React.FC<Props> = (props) => {
 
     const allIds: string[] = [];
     layers.forEach((layer) => {
-      const { pointLayerId, lineLayerId, fillLayerId } = mainManager.getLocationsLayerIds(
+      const { pointLayerId, lineLayerId, fillLayerId } = factoryService.getLocationsLayerIds(
         layer.datasourceId,
         layer.id
       );
@@ -298,7 +308,7 @@ const MainMap: React.FC<Props> = (props) => {
     const locationsByCollection = groupLocationIdsByLayer(locations);
     layers.forEach((layer) => {
       const locationIds = locationsByCollection[layer.id] ?? [];
-      const { pointLayerId, lineLayerId } = mainManager.getLocationsLayerIds(
+      const { pointLayerId, lineLayerId } = factoryService.getLocationsLayerIds(
         layer.datasourceId,
         layer.id
       );
