+++ ui/src/features/Panel/Layers/Layer/index.tsx
@@ -15,8 +15,8 @@ import { useDebounce } from '@/hooks/useDebounce';
 import { useLayerValidation } from '@/hooks/useLayerValidation';
 import { useLoading } from '@/hooks/useLoading';
 import loadingManager from '@/managers/Loading.init';
-import mainManager from '@/managers/Main.init';
 import notificationManager from '@/managers/Notification.init';
+import { collectionService, mainManager } from '@/services/init';
 import { Layer as LayerType } from '@/stores/main/types';
 import { LoadingType, NotificationVariant } from '@/stores/session/types';
 import { CollectionType, getCollectionType } from '@/utils/collection';
@@ -72,7 +72,7 @@ const Layer: React.FC<Props> = (props) => {
       return;
     }
 
-    const collection = mainManager.getDatasource(layer.datasourceId);
+    const collection = collectionService.getDatasource(layer.datasourceId);
 
     if (collection) {
       const collectionType = getCollectionType(collection);
