+++ ui/src/features/Panel/index.tsx
@@ -30,8 +30,8 @@ import styles from '@/features/Panel/Panel.module.css';
 import { Toggle } from '@/features/Panel/Toggle';
 import { Mobile } from '@/features/TopBar/Mobile';
 import loadingManager from '@/managers/Loading.init';
-import mainManager from '@/managers/Main.init';
 import notificationManager from '@/managers/Notification.init';
+import { mainManager } from '@/services/init';
 import useMainStore from '@/stores/main';
 import { LoadingType, NotificationVariant } from '@/stores/session/types';
