+++ ui/src/features/TopBar/Links/GridsChart.tsx
@@ -14,7 +14,6 @@ import { StringIdentifierCollections } from '@/consts/collections';
 import { Charts } from '@/features/Charts';
 import { Parameter } from '@/features/Popup';
 import styles from '@/features/TopBar/Links/Links.module.css';
-import mainManager from '@/managers/Main.init';
 import notificationManager from '@/managers/Notification.init';
 import { TZipLink, ZipService } from '@/services/csv.service';
 import {
@@ -23,6 +22,7 @@ import {
   ICollection,
   IGetCubeParams,
 } from '@/services/edr.service';
+import { collectionService } from '@/services/init';
 import awoService from '@/services/init/awo.init';
 import { Layer } from '@/stores/main/types';
 import { NotificationVariant } from '@/stores/session/types';
@@ -208,7 +208,7 @@ export const GridsChart: React.FC<Props> = (props) => {
   };
 
   useEffect(() => {
-    const collection = mainManager.getDatasource(layer.datasourceId);
+    const collection = collectionService.getDatasource(layer.datasourceId);
 
     if (collection) {
       const paramObjects = Object.values(collection?.parameter_names ?? {});
