+++ ui/src/hooks/useLayerValidation.ts
@@ -9,7 +9,7 @@ import { useEffect, useState } from 'react';
 import { ComboboxData } from '@mantine/core';
 import { CollectionRestrictions, RestrictionType } from '@/consts/collections';
 import { useLoading } from '@/hooks/useLoading';
-import mainManager from '@/managers/Main.init';
+import { collectionService } from '@/services/init';
 import { Layer } from '@/stores/main/types';
 import { CollectionType } from '@/utils/collection';
 import { isSamePalette, isValidPalette } from '@/utils/colors';
@@ -82,7 +82,7 @@ export const useLayerValidation = (layer: Layer, isLoading: boolean, data: Data)
       return;
     }
 
-    const collection = mainManager.getDatasource(layer.datasourceId);
+    const collection = collectionService.getDatasource(layer.datasourceId);
 
     if (collection) {
       const temporalExtent = getTemporalExtent(collection);
