+++ ui/src/hooks/useLocations.ts
@@ -7,8 +7,8 @@ import { useEffect, useRef, useState } from 'react';
 import { Feature } from 'geojson';
 import { StringIdentifierCollections } from '@/consts/collections';
 import loadingManager from '@/managers/Loading.init';
-import mainManager from '@/managers/Main.init';
 import { ICollection } from '@/services/edr.service';
+import { mainManager } from '@/services/init';
 import useMainStore from '@/stores/main';
 import { Layer, Location } from '@/stores/main/types';
 import { LoadingType } from '@/stores/session/types';
