+++ ui/src/hooks/useSpatialSelection.ts
@@ -11,8 +11,8 @@ import { getBBox } from '@/consts/bbox';
 import { LayerId } from '@/features/Map/config';
 import { SourceId } from '@/features/Map/sources';
 import loadingManager from '@/managers/Loading.init';
-import mainManager from '@/managers/Main.init';
 import notificationManager from '@/managers/Notification.init';
+import { mainManager } from '@/services/init';
 import geoconnexService from '@/services/init/geoconnex.init';
 import useMainStore from '@/stores/main';
 import { isSpatialSelectionPredefined } from '@/stores/main/slices/spatialSelection';
